Add files via upload
Hi Everyone,
I have uploaded a copy of my scripts for the 3 questions.  I have provided 2 versions for Q2 2B includes error checking on the file structure for the .csv file locations (as it's pretty specific) and 2A does not include this.
Let me know what you think.  I will also upload to Teams when I can as mine is not functioning tonight.